+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * [FIXED] - #231 - found a problem with message ordering when filtering pings on reconnect, caused issues with reconnect in general
 * [FIXED] - #226 - added more doc about ping intervals and max ping
 * [FIXED] - #224 - resolved a latency problem with windows due to the cost of the message queues spinwait/lock
+* [CHANGED] - started support for renaming gnatsd to nats-server, full release isn't done so using gnatsd for tests still
 
 ## Version 2.4.3